move some code from Navigator to a dedicated class BasicAuthUrl
